Add docstrings & refactor template package.
