+++ pkg/template/doc.go
@@ -1,0 +1,2 @@
+// Package template contains functions for rendering templates.
+package template
